pyproject.toml
[project]
name = "secure-file-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "black>=25.11.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.121.2",
    "flake8>=7.3.0",
    "hiredis>=3.3.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "minio>=7.2.18",
    "mypy>=1.18.2",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
]
